WIP: attempting to have database setup already
I added this command to start the database so that the standard database
setup commands that run when it first starts up are all run in the base
image that we use.  I think it increases the size of the image by a good bit,
but I think it speeds up starting the image.  I have not gotten this
working yet, though.